DERBY-47 (partial): Add some additional IN-list test cases to the lang/inbetween.sql test.  These test cases all currently behave correctly; by adding them to inbetween.sql we can ensure that they will continue to behave correctly once the DERBY-47 changes have been completed.
The underlying notion here is to make sure IN list behavior is
correct when the left operand is a column reference that is a
leading column in one or more indexes.  The DERBY-47 changes
will ultimately make it so that most of the new test cases
result in an index-probing execution plan, thus we want to make
sure that we're testing as many of the various index-based use
cases as possible.

Note that these test cases are just testing correctness of results;
additional tests will be added later to verify that indexes are in
fact being chosen as a result of the DERBY-47 changes.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@512534 13f79535-47bb-0310-9956-ffa450edef68