DERBY-2304 (partial)  Convert derbynet/callable.java to JUnit.
Tests now run in both embedded and client. The suite explicitly runs four
tests that use DriverManager or BigDecimal only if JDBC2 is available.

All tests in derbynet/callable.java were converted except for testBatch, 
which is commented out in the original with the note "Temporarily take out 
testbatch until jcc bug is fixed (5827)". I'll do a little research to see 
if it should now be expected to run successfully.

The test that outputs a message to System.out is not currently enabled
because I'm not sure it makes sense to run it in the junit environment. 
But it would be easily enabled by changing the method name from 
norun_testSystemOutPrintlnProc to testSystemOutPrintlnProc.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@512905 13f79535-47bb-0310-9956-ffa450edef68