DERBY-2304 (partial)  Convert derbynet/callable.java to JUnit.
The old testBatch, which was commented out, is replaced by two new junit tests: 

  - xtestBatchUpdate() batches up calls to a SQL procedure that updates a
    row in a table.
  - xtestBatchUpdateError() does what the first test does, only now one of
    the commands in the batch tries to update the row to a value that violates 
    the constraint check.

The old test is mildly complicated, with a two table schema and primary/foreign
keys. It looked to me like the test had two goals: test a batch that includes 
calls to a sql procedure, and test a batch that does an sql update that 
violates a constraint. (Incidently, the sql update seemed out of place in a 
file that otherwise tests sql procedures and functions.)

I simplified the test for the junit conversion, and it now uses a single table 
with a check constraint on column to test what happens when the constraint is 
violated.  Also the constraint violation is split into a separate test and 
tested in the context of a sql procedure call (in keeping with the rest of 
the tests).  

Left to do: remove derbynet/callable.java


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@514043 13f79535-47bb-0310-9956-ffa450edef68