Patch for DERBY-2416 : Provide a shell for a subclass of SQLChar which will use the passed Collator to do the collation rather than SQLChar's default collation of UCS_BASIC
This patch introduces a new class called CollatorSQLChar which extends SQLChar. The new class has a RuleBasedCollator associated with it
and that Collator object determines the collation. A new format id is added for this class and code regarding format id has gone into 
RegisteredFormatIds, StoredFormatIds and DTSClassInfo. 

The new class is bare minimum at this point and new collation related methods will be added to this class in later patches. This new class
does not get used in Derby at this point. The query compiler code will need to be changed so that this class can get some action. This will 
happen in upcoming patches.

Dan had some suggestion on creating collation elements as required rather than doing it all upfront. Dag also had suggestion on alternative
implementation for array handling involved with collation element creation. We agreed on tackling these issues once the basic functionality 
is in for language based ordering.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@516864 13f79535-47bb-0310-9956-ffa450edef68