DERBY-47 (partial): Patch which creates a new JUnit test based on the repro program attached to the Jira.  The test creates the same kind of table and data as that repro, and then runs three types of queries with larger and larger IN lists. The three types of queries are:
  1 - "Markers" : same as in the repro program
  2 - "Literals" : same as in the repro program
  3 - 'MixedIds": IN list has a combination of parameter markers and literals.

For each query we check to make sure the results are correct and then we look
at the query plan to determine whether or not the optimizer chose to do multi-
probing. If the results are incorrect or if the optimizer did *not* choose
multi-probing then the test will fail.

The test determines that "multi-probing" was in effect by looking at the query
plan and verifying two things:

  1. We did an index scan on the target table AND

  2. The number of rows that "qualified" is equal to the number of rows that
     were actually returned for the query.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@520191 13f79535-47bb-0310-9956-ffa450edef68