Committing patch DERBY2535_Return_Collator_api_On_DVD_v1_diff.txt attached to DERBY2535. This patch adds the api on DVF to return the Collator object depending on the collation type. The new api looks as follows RuleBasedCollator getCharacterCollator(int collationType);
If the collation type is UCS_BASIC, then Collator object returned by the new api above will be null. If the collation type is 
TERRITORY_BASED, then Collator object returned will be the one based on the territory of the database. 

This new api will be used in a method on DataValueFactory(DVF) which will take format id and collation type and return a DVD based on 
those 2 values. That method on DVF will first consturct a DVD using just the format id. For character trype format ids, this means that 
the DVF will always be SQLChar/SQLVarchar/SQLLongvarchar/SQLClob. This is ok if we are dealing with collation type of UCS_BASIC. But if 
the collation type is TERRITORY_BASED, then we shoould construct a DVD of type 
CollatorSQLChar/CollatorSQLVarchar/CollatorSQLLongvarchar/CollatorSQLClob. In order to construct these collation sensitive DVDs, the method 
on DVF will call the new api on DVF (getCharacterCollator) and it will use the return Collator from getCharacterCollator to construct the 
right kind of character DVD. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@528070 13f79535-47bb-0310-9956-ffa450edef68