DERBY-2557 : This commit renames getInstanceUsingFormatIdAndCollationType method on DVF to getNull. It also changes the implementation of the that method so that it bypasses the InstanceGetter. This implementation is much simpler than the old one because there is no InstanceGetter involved. Lastly, the code associated with getting a DVD from a format id is moved out of DTSClassinfo into a static method on DataValueFactoryImpl. This static method will be called by DTSClassinfo.
The changes made in the patch are as follows 
1)DataValueFactory 
Changed the name of the new interface from getInstanceUsingFormatIdAndCollationType to getNull. This interface now returns a DVD rather 
than an Object. The functionality of the interface remains the same but the implementation has changed. 
2)DataValueFactoryImpl 
a)Removed the class level array and the code associated with InstanceGetter. 
b)Added a new static method called getNullDVDWithUCS_BASICcollation(int formatId). This static method handles all the format ids associated 
with DVDs with the exception of Decimals. The class to be returned for Decimals depends on what VM is being used. This dependency on the VM 
is handled by getNullDecimal defined on DVF. But since getNullDecimal is not a static method, it can't be called by static method 
getNullDVDWithUCS_BASICcollation. I could go the path of defining getNullDecimal as static but that will require changes in 
NumericTypeCompiler.nullMethodName method. Also, all the other getNullXXX on DVF are non-static So, the code for returning the right DVD 
for Decimal is not in getNullDVDWithUCS_BASICcollation. Rather it is in the calling method, getNull. For other format ids associated with 
DVDs, getNull will check if the DVD is of type StringDataValue and the collation type is territory based and if so, then it will return 
((StringDataValue)returnDVD).getValue(getCharacterCollator(collationType)); 
3)DTSClassInfo 
This class now calls the static method in DVF to get the DVDs. But if the format id is not for a DVD, then it checks if it needs to return 
TypeId. 


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@530546 13f79535-47bb-0310-9956-ffa450edef68