DERBY-2598
Fixes upgrade bug in writeExternal of Heap, it was writing the wrong 
conglomerate format id in the case of a hard upgrade accessing an 
existing Heap from an old database.  It was writing out the hardcoded
new format
id when it should have been writing out the dynamic one in the class's
state.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@535521 13f79535-47bb-0310-9956-ffa450edef68