DERBY-1054: NetServlet does not allow binding to non localhost interface
This change modifies the NetServlet.java class to allow for a "host"
parameter to be passed in via the web.xml file. That way we can control
the listening interface. The sample web.xml is enhanced with additional
information about how to set the host parameter.

Note that there are security implications to setting a non-localhost
value for the host. Such a server should be properly configured so that
access to it is authenticated and secured. A separate docs patch enhances
the NetServlet documentation to note these considerations.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@535910 13f79535-47bb-0310-9956-ffa450edef68