DERBY-2019: IJ describe command does not handle quotes very well
IJ's describe command already accepted a string for the table name,
but there were several problems with the handling of the table name
in single quotes:
1) The table name was unnecessarily forced to upper case.
2) The command only worked properly if you gave both schema and table name.
3) It had a rather bizarre behavior with an empty string

This patch removes the force-to-upper, cleans up the schema name handling,
addresses the empty string case by treating it as an error,
adds a simple '*' feature as a wildcard, and adds some new tests.

With the patch,
   describe 'CamelCaseTable';
can be used to describe a table with a mixed-case name. 

A separate patch will update the documentation to describe the behavior.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@536227 13f79535-47bb-0310-9956-ffa450edef68