DERBY-2141
Made BaseContainerHandle.close() synchronized as it may be called concurrently
by a stream passed outside the store and an abort triggered concurrently.  
Making it synchronized will allow the first close to do the proper work and
the subsequent calls will just return with no work to do.
Also made some of the fields private to convince myself no-one outside the
class was counting on something being changed during close.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@537091 13f79535-47bb-0310-9956-ffa450edef68