DERBY-2597: Language result sets should not reuse current isolation level across executions
1) Move the static method UpdateResultSet.decodeLockMode() into
DMLWriteResultSet and make it non-static. This makes it possible to
add debug code that checks that it is only called when the
{Update,Delete,Insert}ResultSet has been opened (to prevent calls at
ResultSet construction time).

2) Move UpdateResultSet's call to decodeLockMode() from the
constructor to open().

3) Make the NoRowsResultSetImpl class hierarchy maintain an isOpen
variable that is set to true on open and to false on
close. Previously, an isClosed variable was used, which was false
until the first call to close(), after which it would be true forever
even after the RS was reopened. This change was necessary to be able
to assert that the result set had been opened in decodeLockMode().


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@537128 13f79535-47bb-0310-9956-ffa450edef68