DERBY-2537
Check in intermediate version of a new test to test changes associated with
this issue.  There is still more test cases to add.  It is not ready to 
be included in the all suite.  It passes after the 
DERBY2335_Use_PushDataValueFactory_in_ExpressionClassBuilder_v1_diff.txt 
patch of DERBY-2335 is applied - otherwise fails with Class cast exceptions.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@538300 13f79535-47bb-0310-9956-ffa450edef68