DERBY-2558: Client throws JDK exception rather than parameter out of range
The client implementation throws an ArrayIndexOutOfBoundsException rather
than the expected: ERROR XCL13: The parameter position '2' is out of range.
The number of parameters for this prepared statement is '1'. 

The issue is that the 4-argument overload of setObject was missing the
code to validate the parameterIndex; this code is present in the other
overloads of setObject.

This change adds the parameterIndex validation call, and also adds tests.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@538413 13f79535-47bb-0310-9956-ffa450edef68