DERBY-2674 (partial) Stop running jdbc40 tests in the old test framework
The attached patch removes the TestPreparedStatementMethods test and
associated files that are not used by other tests.

An inspection of PreparedStatementTest showed that testing of
isPoolable/setPoolable on a closed PreparedStatement was omitted when
the test was ported to Junit. The patch adds test cases for this to
PreparedStatementTest.

Contributed by Øystein Grøvlen.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@540150 13f79535-47bb-0310-9956-ffa450edef68