DERBY-1816 (partial): Pre-patch "cleanup" that does the following:
  1) Replaces each of the recyclable Date, Time, and Timestamp arguments
     with a recyclable java.util.Calendar object in client/am/Cursor.java.

  2) Modifies the relevant code in client/am/DateTime.java to call methods
     on the recyclable Calendar object instead of on Date, Time, and
     Timestamp objects. The benefit to doing this is that we are now using
     non-deprecated methods.

Note that even with this patch we are still creating a new instance of
Time/Timestamp/Date for each method--the cleanup patch does not change that.
Instead, the cleanup patch adds the instantiation of a new Calendar object
(one per client/am/Cursor) and then (re-)uses that object to replace the
deprecated calls.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@540740 13f79535-47bb-0310-9956-ffa450edef68