DERBY-2776 These changes already commited to main codeline. Now, commiting them to 10.3 codeline.
Internally generated CAST nodes should not pick up the collation of the current schema. In order to implement this, the CAST nodes generated directly by the user sql (parser) will set a flag on the cast node to indicate that they are externally generated CAST nodes. During the bind phase of a CAST node, we will check if the node is externally generated. If yes, then we will have it pick up the collation of the compilation schema otherwise we will leave the collation unchanged. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.3@551485 13f79535-47bb-0310-9956-ffa450edef68