DERBY-2831 Merging changes from main into 10.3 codeline. The merge command is as follows svn merge -r 551792:551793  https://svn.apache.org/repos/asf/db/derby/code/trunk/
The commit comments for 10.3 codeline were as follows
DERBY-2831 Creating a function that returns a character string literal inside a schema that doesnot exist should not give 
a null pointer exception. This was happening while we were trying to look up the collation. The collation lookup should be 
done using the super class's getSchemaDescriptor method. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.3@551794 13f79535-47bb-0310-9956-ffa450edef68