DERBY-2878 (partial) Scan protection handle could be cached in BasePage
This patch is the first step towards eliminating the allocation of
protection handles. It adds a method to Page/BasePage called
getProtectionRecordHandle(). This method creates a new record handle
if it hasn't been called before and returns a cached value
otherwise. B2IRowLocking3.lockScan() and lockScanForReclaimSpace()
have been modified to use the new method.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@551933 13f79535-47bb-0310-9956-ffa450edef68