Merging change 552531 from main into 10.3.1.1 codeline. The commit comments for the main were as follows
DERBY-2879 

Derby currently requires all the character columns in a table to have the same collation as the collation of the schema in 
which the table is getting defined. In order to implement this behavior, CreateTableNods's bind method will check the 
collation of it's character columns against the schema's collation and if there is a mismatch, an exception will be thrown. 

In addition, this patch does minor cleanup. The string constants for the 2 collation types in Derby were declared in both 
StringDataValue.java and Property.java I have removed it from StringDataValue so we don't have duplicate constants. 


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.3@552534 13f79535-47bb-0310-9956-ffa450edef68