Merging change 553395 from main codeline for DERBY-2793. The commit comments for that change were as follows
This method fixes LikeEscapeOperatorNode so that it will check the collation derivation too rather than only checking the
collation type. This fix will make sure that if one side of the LIKE operation has collation derivation of NONE and the
other side has collation derivation of IMPLICIT, then we should throw an exception.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.3@553399 13f79535-47bb-0310-9956-ffa450edef68