DERBY-2777
A trivial change to CollationTest2 test. Rather than relying on the number of rows in SYSCOLUMNS for the test validity,
I am changing it so that no rows get selected from the LIKE test. This is so that over the time, if we end up changing
the SYSCOLUMNS so that it will have more rows than what we have in the codeline today, we won't have to go and fix the
test. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@553703 13f79535-47bb-0310-9956-ffa450edef68