DERBY-2777
Currently, the parameters in LOCATE clause always pickup their collation from the compilation schema. That logic is not 
complete. I am fixing that logic here along with addition of some tests. 

For the sake of explanation, let me use the following syntax for LOCATE clause 
LOCATE (receiver, leftOperand) 
With the fix in this patch, if receiver is a parameter, it will set it's collation using following logic 
1)check if leftOperand is not a parameter. If yes, then receiver will pick up collation from leftOperand. If not, goto step 2 
2)receiver picks up the collation of the compilation schema because everything in the LOCATE clause is ? 

Next, if leftOperand is a parameter, it will set it's collation using receiver. By this time, even if receiver is a 
parameter, we have set correct collation for receiver and hence leftOperand can simply rely on receiver for correct 
collation IF leftOperand is a parameter. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@553784 13f79535-47bb-0310-9956-ffa450edef68