DERBY-2902: can't use Long.MIN_VALUE as the start value for a generated column
When creating a table
    AS IDENTITY (START WITH -9223372036854775808)
fails but
    AS IDENTITY (START WITH -9223372036854775807)
succeeds.

The problem was that the parser was handling the +/- sign separately
from the start value, and was trying to call
Long.parseLong("9223372036854775808") and then negate the return value,
which fails because 9223372036854775808 is greater than Long.MAX_VALUE.

The fix is to push the handling of the "-" sign into the call to
Long.parseLong(), where it is handled correctly. We have to be careful
not to push the "+" sign into parseLong(), though, because it doesn't
accept a leading "+" sign, only a leading "-" sign.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@554996 13f79535-47bb-0310-9956-ffa450edef68